third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/subresourceIntegrity/subresource-integrity-style-no-cors-no-xorigin-with-creds-expected.txt

CONSOLE ERROR: Subresource Integrity: The resource 'http://localhost:8000/security/resources/cors-style.php?cors=false&credentials=true' has an integrity attribute, but the resource requires the request to be CORS enabled to check the integrity, and it is not. The resource has been blocked because the integrity cannot be enforced.
This is a testharness.js-based test.
PASS SRI on cross origin CORS disabled but with credentials style, without CORS fetch 
Harness: the test ran to completion.
